Presentation Draft
Almost all of it is done. The part about the game we are making is not
finished. I figured it might be best for you to add what you feel is
correct. Also I was not sure if we should cover a brief history of video
games or RPG's in general. I hope you like it. Cheers