Midterm Report
The only person not in the report yet is David. I will get him to add it
Monday later on. Please revise it and make sure it looks good and is
correct. If anyone has a problem with what another group member did or
did not write, contact them so we can get this squared away.